Fix for debug printf format strings. size_t are not expected.
git-svn-id: https://pupnp.svn.sourceforge.net/svnroot/pupnp/trunk@181 119443c7-1b9e-41f8-b6fc-b9c35fce742c

+++ upnp/src/gena/gena_device.c
@@ -205,7 +205,7 @@ notify_send_and_recv( IN uri_type * destination_url,
     // connect
     DBGONLY( UpnpPrintf( UPNP_ALL, GENA, __FILE__, __LINE__,
                          "gena notify to: %.*s\n",
-                         destination_url->hostport.text.size,
+                         (int)destination_url->hostport.text.size,
                          destination_url->hostport.text.buff ); )
 
         conn_fd = http_Connect( destination_url, &url );
